+++ packages/derivatives-consumer/src/transformers/DerivativeTransformer.ts
@@ -2,23 +2,28 @@ import { TradeDirection, TradeExecutionType } from '@injectivelabs/ts-types'
 import {
   GrpcPriceLevel,
   GrpcDerivativeMarketInfo,
-  GrpcDerivativeMarketOrder,
+  GrpcDerivativeLimitOrder,
   GrpcDerivativeTrade,
   DerivativeOrderType,
   Orderbook,
   PriceLevel,
   DerivativeMarket,
-  DerivativeMarketOrder,
+  DerivativeLimitOrder,
   DerivativeMarketTrade,
   DerivativeOrderState,
   GrpcTokenMeta,
   TokenMeta,
+  PositionDelta,
+  GrpcDerivativePosition,
+  Position,
+  GrpcPositionDelta,
 } from '../types'
 
-const zeroPriceLevel = () => ({
-  price: '0',
-  quantity: '0',
-  timestamp: 0,
+const zeroPositionDelta = () => ({
+  tradeDirection: TradeDirection.Buy,
+  executionPrice: '0',
+  executionQuantity: '0',
+  executionMargin: '0',
 })
 
 export class DerivativeTransformer {
@@ -43,16 +48,18 @@ export class DerivativeTransformer {
     market: GrpcDerivativeMarketInfo,
   ): DerivativeMarket {
     return {
+      oracleBase: market.getOracleBase(),
+      oracleQuote: market.getOracleQuote(),
+      oracleType: market.getOracleType(),
+      initialMarginRatio: market.getInitialMarginRatio(),
+      maintenanceMarginRatio: market.getMaintenanceMarginRatio(),
+      isPerpetual: market.getIsPerpetual(),
       marketId: market.getMarketId(),
       marketStatus: market.getMarketStatus(),
       ticker: market.getTicker(),
-      baseDenom: market.getBaseDenom(),
       quoteDenom: market.getQuoteDenom(),
       quoteToken: DerivativeTransformer.grpcTokenMetaToTokenMeta(
         market.getQuoteTokenMeta(),
-      ),
-      baseToken: DerivativeTransformer.grpcTokenMetaToTokenMeta(
-        market.getBaseTokenMeta(),
       ),
       makerFeeRate: market.getMakerFeeRate(),
       takerFeeRate: market.getTakerFeeRate(),
@@ -68,6 +75,17 @@ export class DerivativeTransformer {
     return markets.map((market) =>
       DerivativeTransformer.grpcMarketToMarket(market),
     )
+  }
+
+  static grpcPositionDeltaToPositionDelta(
+    positionDelta: GrpcPositionDelta,
+  ): PositionDelta {
+    return {
+      tradeDirection: positionDelta.getTradeDirection() as TradeDirection,
+      executionPrice: positionDelta.getExecutionPrice(),
+      executionQuantity: positionDelta.getExecutionQuantity(),
+      executionMargin: positionDelta.getExecutionMargin(),
+    }
   }
 
   static grpcPriceLevelToPriceLevel(priceLevel: GrpcPriceLevel): PriceLevel {
@@ -100,33 +118,59 @@ export class DerivativeTransformer {
   }
 
   static grpcOrderToOrder(
-    order: GrpcDerivativeMarketOrder,
-  ): DerivativeMarketOrder {
+    order: GrpcDerivativeLimitOrder,
+  ): DerivativeLimitOrder {
     return {
       orderHash: order.getOrderHash(),
       orderType: order.getOrderType() as DerivativeOrderType,
       marketId: order.getMarketId(),
       subaccountId: order.getSubaccountId(),
+      isReduceOnly: order.getIsReduceOnly(),
+      margin: order.getMargin(),
       price: order.getPrice(),
-      state: order.getState() as DerivativeOrderState,
       quantity: order.getQuantity(),
       unfilledQuantity: order.getUnfilledQuantity(),
       triggerPrice: order.getTriggerPrice(),
       feeRecipient: order.getFeeRecipient(),
+      state: order.getState() as DerivativeOrderState,
     }
   }
 
   static grpcOrdersToOrders(
-    orders: GrpcDerivativeMarketOrder[],
-  ): DerivativeMarketOrder[] {
+    orders: GrpcDerivativeLimitOrder[],
+  ): DerivativeLimitOrder[] {
     return orders.map((order) => DerivativeTransformer.grpcOrderToOrder(order))
   }
 
+  static grpcPositionToPosition(order: GrpcDerivativePosition): Position {
+    return {
+      marketId: order.getMarketId(),
+      subaccountId: order.getSubaccountId(),
+      direction: order.getDirection() as TradeDirection,
+      quantity: order.getQuantity(),
+      entryPrice: order.getEntryPrice(),
+      margin: order.getMargin(),
+      holdQuantity: order.getHoldQuantity(),
+      liquidationPrice: order.getLiquidationPrice(),
+      markPrice: order.getMarkPrice(),
+      impliedPnl: order.getImpliedPnl(),
+      leverage: order.getLeverage(),
+    }
+  }
+
+  static grpcPositionsToPositions(
+    positions: GrpcDerivativePosition[],
+  ): Position[] {
+    return positions.map((position) =>
+      DerivativeTransformer.grpcPositionToPosition(position),
+    )
+  }
+
   static grpcTradeToTrade(trade: GrpcDerivativeTrade): DerivativeMarketTrade {
-    const price = trade.getPrice()
-    const mappedPrice = price
-      ? DerivativeTransformer.grpcPriceLevelToPriceLevel(price)
-      : zeroPriceLevel()
+    const positionDelta = trade.getPositionDelta()
+    const mappedPositionDelta = positionDelta
+      ? DerivativeTransformer.grpcPositionDeltaToPositionDelta(positionDelta)
+      : zeroPositionDelta()
 
     return {
       orderHash: trade.getOrderHash(),
@@ -134,9 +178,10 @@ export class DerivativeTransformer {
       marketId: trade.getMarketId(),
       executedAt: trade.getExecutedAt(),
       tradeExecutionType: trade.getTradeExecutionType() as TradeExecutionType,
-      tradeDirection: trade.getTradeDirection() as TradeDirection,
       fee: trade.getFee(),
-      ...mappedPrice,
+      isLiquidation: trade.getIsLiquidation(),
+      payout: trade.getPayout(),
+      ...mappedPositionDelta,
     }
   }